Move OS specific functions to utilos.h/cpp
Limits the warning output of Carbon.h
