+++ Client/qtTeamTalk/CMakeLists.txt
@@ -91,6 +91,7 @@ if (Qt5_FOUND OR Qt6_FOUND)
     utilsound.h utilvideo.h utiltts.h utilui.h utilhotkey.h
     serverlogeventsmodel.h mytreeview.h textmessagecontainer.h
     useraccountsmodel.h encryptionsetupdlg.h utiltt.h utilxml.h
+    utilxml.h
 
     main.cpp mainwindow.cpp preferencesdlg.cpp uservideowidget.cpp
     channelstree.cpp channeldlg.cpp userinfodlg.cpp
@@ -110,6 +111,7 @@ if (Qt5_FOUND OR Qt6_FOUND)
     utilsound.cpp utilvideo.cpp utiltts.cpp utilui.cpp utilhotkey.cpp
     serverlogeventsmodel.cpp mytreeview.cpp textmessagecontainer.cpp
     useraccountsmodel.cpp encryptionsetupdlg.cpp utiltt.cpp utilxml.cpp
+    utilos.cpp
 
     mainwindow.ui channel.ui preferences.ui
     serverlist.ui userinfo.ui bannedusers.ui useraccounts.ui
