+++ Client/qtTeamTalk/mainwindow.pri
@@ -22,7 +22,8 @@ HEADERS    = mainwindow.h preferencesdlg.h uservideowidget.h \
              statusbardlg.h statusbareventsmodel.h mycombobox.h \
              utilsound.h utilvideo.h utiltts.h utilui.h utilhotkey.h \
              serverlogeventsmodel.h mytreeview.h textmessagecontainer.h \
-             useraccountsmodel.h encryptionsetupdlg.h utiltt.h utilxml.h
+             useraccountsmodel.h encryptionsetupdlg.h utiltt.h utilxml.h \
+             utilos.h
 
 SOURCES    = main.cpp mainwindow.cpp preferencesdlg.cpp uservideowidget.cpp \
              channelstree.cpp channeldlg.cpp userinfodlg.cpp \
@@ -41,7 +42,8 @@ SOURCES    = main.cpp mainwindow.cpp preferencesdlg.cpp uservideowidget.cpp \
              statusbardlg.cpp statusbareventsmodel.cpp mycombobox.cpp \
              utilsound.cpp utilvideo.cpp utiltts.cpp utilui.cpp utilhotkey.cpp \
              serverlogeventsmodel.cpp mytreeview.cpp textmessagecontainer.cpp \
-             useraccountsmodel.cpp encryptionsetupdlg.cpp utiltt.cpp utilxml.cpp
+             useraccountsmodel.cpp encryptionsetupdlg.cpp utiltt.cpp utilxml.cpp \
+             utilos.cpp
 
 FORMS      = mainwindow.ui channel.ui preferences.ui \
              serverlist.ui userinfo.ui bannedusers.ui useraccounts.ui \
